Make new asynchronous google analytics optional.
The Asynchronous tracking snippet is not exactly a drop it replacement for the old
synchronous tracking snippet.  It is intended to go at a different place in the template,
and it doesn't work with some documents (see README.mdown for details).  Therefore, it
makes sense for it to be rendered using a different template tag.  It could have been an
option passed to the template tag, but template tags should be simple to use, and simple
to parse.